fix types and fallback (#210)
